job #8006 - Got rid of the independent function gen plugin.  Moved the functionality into the bp.utilities plugin along with our other small tools.
