--- src/org.xtuml.bp.function.gen/.gitignore
@@ -1,1 +1,0 @@
-/bin/
