--- src/org.xtuml.bp.function.gen/build.properties
@@ -1,6 +1,0 @@
-source.. = src/
-output.. = bin/
-bin.includes = META-INF/,\
-               .,\
-               plugin.xml,\
-               about.html
