--- src/org.xtuml.bp.function.gen/plugin.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?eclipse version="3.4"?>
-<plugin>
-   <extension
-         point="org.eclipse.ui.popupMenus">
-      <objectContribution
-            id="org.xtuml.bp.function.gen.contribution1"
-            objectClass="org.xtuml.bp.core.Package_c">
-         <action
-               class="org.xtuml.bp.function.gen.action.GenerateFunctionsFromList"
-               enablesFor="1"
-               id="org.xtuml.bp.function.gen.newAction"
-               label="Generate Functions From List..."
-               menubarPath="org.xtuml.bp.utilities/org.xtuml.bp.utilities.separator">
-         </action>
-      </objectContribution>
-   </extension>
-
-</plugin>
